# src/main/resources/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "alojamiento quindio",
    "description": "API REST para la gestión completa",
    "contact": {
      "name": "Equipo de Desarrollo",
      "url": "https://tiendavirtual.com",
      "email": "dev@tiendavirtual.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor de Desarrollo"
    },
    {
      "url": "https://api-test.tiendavirtual.com",
      "description": "Servidor de Pruebas"
    },
    {
      "url": "https://api.tiendavirtual.com",
      "description": "Servidor de Producción"
    }
  ],
...
  "paths": {
...
    "/api/v1/guest/bookings/{id}": {
      "put": {
        "tags": [
          "Guest Bookings"
        ],
        "summary": "Modificar reserva",
        "description": "Solo para huésped propietario; valida políticas de modificación.",
        "operationId": "updateBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la reserva",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingUpdateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Reserva modificada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "400": {
            "description": "Validación fallida",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "403": {
            "description": "No propietario",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Guest Bookings"
        ],
        "summary": "Cancelar reserva",
        "description": "Solo para huésped propietario; según políticas de cancelación, genera notificación.",
        "operationId": "cancelBooking",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la reserva",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        ],
        "responses": {
          "204": {
            "description": "Reserva cancelada"
          },
          "400": {
            "description": "Fuera de plazo de cancelación"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "No propietario"
          },
          "404": {
            "description": "No encontrado"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/guest/bookings": {
      "get": {
        "tags": [
          "Guest Bookings"
        ],
        "summary": "Listar reservas del huésped",
        "description": "Historial con filtros (activas, canceladas). Paginado.",
        "operationId": "getGuestBookings",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Estado de la reserva",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "CONFIRMED"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Número de página",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Tamaño de página",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de reservas",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingDTO"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Guest Bookings"
        ],
        "summary": "Crear reserva",
        "description": "Valida disponibilidad; genera notificación a host.",
        "operationId": "createBooking",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreateDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Reserva creada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "400": {
            "description": "No disponible o validación fallida",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          },
          "403": {
            "description": "Rol no autorizado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BookingDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/guest/dashboard": {
      "get": {
        "tags": [
          "Guest Dashboard"
        ],
        "summary": "Obtener resumen del dashboard del huésped",
        "description": "Incluye historial resumido, reservas activas, comentarios pendientes y métricas personales.",
        "operationId": "getGuestDashboard",
        "responses": {
          "200": {
            "description": "Resumen obtenido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummaryDTO"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardSummaryDTO"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "BookingDTO": {

      },
      "BookingUpdateDTO": {
        "description": "Datos actualizados de la reserva"
      },
...
      "BookingCreateDTO": {
        "description": "Datos de la reserva a crear"
      },
...
      "DashboardSummaryDTO": {

      },
...
    }
  }
}